moving couple of functions
